requirements.txt
pyfiglet
tkinter
Pillow
opencv-python